pyproject.toml
[project]
name = "vcard-generator"
version = "0.1.0"
description = "vCard (VCF) Generator"
requires-python = ">=3.11"

dependencies = [
    "python-fasthtml",
    "MonsterUI",
    "uvicorn",
    "segno"
]

[tool.setuptools]
py-modules = ["app"]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "vobject",
    "pillow",
    "qreader"
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"